ZMapLocationEdit+ZLocationCell added: MapKit based editor for a location name and a geo coordinate (see "Special" category in the ZDetailKitDemo app)
ZMapLocationEdit is based on ZDetailViewBaseController, as it is a detail view controller
but is NOT using a tableview, but a custom layout.

This shows how simple it is to convert a regular view controller into a detail editor:
- create valueConnectors for the edited data
- override setActive to add code to be executed when editor is being connected to
  or disconnected from data.
- if needed, also override detailViewWillOpen/Close
- done

ZDetailViewBaseController handles all the interfacing with the caller (usually a
cell from another detailview), save/cancel, presentation, navigation buttons etc.