Improved input view again, hopefully now really ok for iPhone and iPad, including rotation
Unlike in previous commit, anchoring the input view in self.view after animating in
is not a good idea, as this means the input view scrolls with the table.
It however solves animating out problems, so the trick is now to re-anchor it
to the scroll view just before starting the out animation.

Furthermore, the search for the view to attach the inputview to now uses the parent view of the first non-UIScrollView, starting from self.view. This seems to get the right view for iPhone and iPad.