ZValueConnectorContainer: contained containers are not a good idea - removed them again.
If there are contained objects with value connectors - there is no simple rule about when and how they should be activated/saved/validated. This is specific to the object that embeds them, and therefore a generalized nested container is not helpful.

For example ZDetailTableViewController is a container for a specific type of object with connectors, namely ZDetailViewCells, with its own semantics about when to enable/save/load the cells's connectors.