ZDetailViewBaseCell: neededModes converted to showInModes (ORed conditions now)
The flags for the display level (basic/detail, empty or not) and for the editing state (viewing/editing, empty or not) are evaluated separately.

If no level flag (ZDetailDisplayModeLevelMask) is present in showInModes, cell shows in all levels.
If no editing state flag (ZDetailDisplayModeEditingMask) is present, cell shows in view and edit modes.

This makes the setting for each cell more readable.