Input view and scroll adjustment: another nasty detail: rootViewController's view might not have a window and hence not work for coordinate calculations (will return wrong values on retina).
This happens when the rootView is not visible because another view was presented fullscreen and modally on top of it.

Fix is to check window==nil, and if so, walk down the presentedViewController chain to find a view which has a window!=nil and use it instead as full screen reference.